Tests and switch some of the bit encoding to use bitwise logic instead. It's cleaner, faster and is less confusing to follow.

+++ src/ins/instruction.rs
@@ -36,7 +36,7 @@ pub enum Instruction {
     MovU32RegPtrU32RegRelSimple(RegisterId, RegisterId),
 
     /******** [Complex Move Instructions - WITH EXPRESSIONS] ********/
-    MovU32ImmMemRelExpr(u32, u16),
+    MovU32ImmMemRelExpr(u32, u32),
 
     /******** [Special Instructions] ********/
     Ret,
